IT WORKS ON SUGAR
