+++ activity/activity.info
@@ -4,7 +4,7 @@ activity_version = 2
 bundle_id = ctcubc.Metris
 icon = activity
 show_launcher = yes
-exec = sugar-activity TestActivity.TestActivity
+exec = sugar-activity metris-activity.MetrisActivity
 repository = https://github.com/CodeTheChangeUBC/OLPC
 license = GPLv3+
 category = maths game
